Ignore Gemfile locks
If we have any explicit version dependency with a library, we should enforce that in the project's gemspec. Otherwise, we could have the illusion our builds are passing when third-party users are having their own deployments with potentially different versions (Gemfile.lock isn't packaged with the gem).

Therefor, I think the wise solution is not to version control it in our repository.