Codecov improvements (#1314)
* Fix codecov configuration

Previous configuration was invalid, so the default codecov config +
whatever team settings we have is what has been used so far. You can
check this with:

```
curl --data-binary @codecov.yml https://codecov.io/validate
```

What I did was to copy over codecov's default configuration because:

* We seem happy with the current settings (nobody has complained).
* We like explicit configuration (the file was there).

The only deviation from this is the required "diff coverage" for the
check to pass. The default is the global coverage percentage of the
project, which at the moment is ~95%. I changed it to 50% because:

* I find ~95% too much for a required check. If any PR changes an
  uncovered line (a style change or simple refactoring, for example), the
  check will most likely fail unless the size of the diff is very big.
* The original file had this already set to 50%. It was just not applied
  because the file was invalid.

* Style change to prove new config works as expected