Attachment size (#1352)
* Remove harcoded maximum image size

This probably needs to be made more flexible, but for now I don't think
we want the limit hardcoded here.

* Typo

* Revert "Evaluate `maximum_attachment_size` at require time (#1315)"

This reverts commit c7a7cb705e39564cf7b55f7b474a726569f6abff. We don't
want this. Since the models are required on boot _before_ initializers
are applied, the limit will be set to the default value before the
custom config gets the chance to be applied. So we indeed need to lazy
evaluate it. Sorry about that :(