Core refactors & improvements (#1377)
* Quote class_name in associations

* It leads to no circular loading issues when using STI or things like
that.
* It's strictly more correct, since a constant is not a class name.
* It normalizes all associations.

* Rename and rewrite query docs to be less confusing

This query has nothing to do with public processes. As a matter of fact,
the `published` filtering is done in the
`OrganizationParticipatoryProcesses` query. So I improved the name and
description of the query.

* Dry up duplicated logic into a query and test it

* Remove PublicProcesses service

The other two helper methods in this controller are implemented either
directly in the controller or as queries. This one is just data
filtering as well, so I don't see why it should be different.

* Normalize nomenclature for engines

Always call regular engines "Engine", and admin engines "AdminEngine".
The fact the two different names for the main engine of plugins confused
my grepping a bit.

* Remove stuff already included in HasAttachments