Don't use hover with poltergeist
Every once in a while, the element won't be properly hovered and the
dropdown menu won't expand. Since the behavior on hover and click is the
same, we can just click. Furthermore, this behaviour is consistent
regardless of screen size or user agent, so it seems better to test
that.