Properly wait for the foundation JS to be ready
Without properly waiting for the plugin to be ready, we get the
following race condition:

* We get to `page.find("a", text: "Random")` before foundation has been
initialized, so the click will do nothing and won't expand the menu.
* Now we go on to `click_link "Recent"` which is a visible link even
without the foundation plugin. The link is properly located by
`click_link` but not yet clicked.
* Now the foundation plugin is finally loaded, so the "Recent" link
disappears.
* Capybara clicks on the coordinates located previously for the "Recent"
link, but what's actually there now is a link to the first proposal
page.

This is my guess at why sometimes the test would fail and the failure
screenshot would be located in the proposal show page.