Fix proposal testing issues (#1407)
* Improve some expectations readability

Before they were a bit too shakespearean...

* Fix bug in dropdown menu on mobile devices

When tapping on the dropdown menu, the menu would appear briefly but the
page would be very quickly reloaded, making the sorting options pretty
unusable.

Use a placeholder link instead of a real link to prevent that.

* Prevent empty orders causing a raw hash label

If the order param is sent empty, it was causing a "raw hash label" to
appear in the UI, something like this:

```ruby
{:label=>"Ordenar propostes per:",:most_voted=>"Més
votat",:random=>"Aleatori",:recent=>"Recent"}
```

* Move `click_link`'s inside `within`

It wouldn't be a problem unless there were other similar links in other
places in the page, very unlikely. But I think it reads nicer this way.

* Don't use hover with poltergeist

Every once in a while, the element won't be properly hovered and the
dropdown menu won't expand. Since the behavior on hover and click is the
same, we can just click. Furthermore, this behaviour is consistent
regardless of screen size or user agent, so it seems better to test
that.

* Properly wait for the foundation JS to be ready

Without properly waiting for the plugin to be ready, we get the
following race condition:

* We get to `page.find("a", text: "Random")` before foundation has been
initialized, so the click will do nothing and won't expand the menu.
* Now we go on to `click_link "Recent"` which is a visible link even
without the foundation plugin. The link is properly located by
`click_link` but not yet clicked.
* Now the foundation plugin is finally loaded, so the "Recent" link
disappears.
* Capybara clicks on the coordinates located previously for the "Recent"
link, but what's actually there now is a link to the first proposal
page.

This is my guess at why sometimes the test would fail and the failure
screenshot would be located in the proposal show page.

* Follow foundation advice on preventing FUOC

This is one of the things I tried when fighting with these bugs. It was
not the real issue, but I guess it doesn't hurt?

http://foundation.zurb.com/sites/docs/dropdown-menu.html#preventing-fouc

* Remove dead code branch

* Consistently pass symbol to method

* Prefer click_link

* Remove unnecessary object from test

* Simplify setting proposal order

* Replace CSS with actual content in expectation

* Prevent some duplicate visits

* Extract ordering stuff to a concern

* Add missing EOF

* Refactor order selection to a single place

This led to another bug fix. Previously when votes were disabled, it
would still allow sorting by "most voted" by direct parameter in the
URL.