Locale sanity (#1330)
* Allow checking i18n against non English locales

* Try out dynamic missing locale guessing

* Move i18n specs to shared examples

* More readable test failures

With colors and all.

* Implement locale checking from installations

* Use a better set of locales in development_app

Currently `rake development_app` generates a development app with all
locales supported by decidim. Since at the moment is compulsory to
introduce all locale information in forms, it's very tedious to make
manual tests using this app. So I changed it to use the same the test
app uses.

* Task to enforce completeness of official locales

As a release pre-requisite.

* Add some docs to the `decidim:check_locales` task

* Fix style issue detected by Codeclimate