Fix default button when no voting enabled (#1429)
* Remove global setup from proposal tests

Most of the tests here don't need all the object creations in the global
setup. So move that only where it's needed. This makes the tests faster,
but more importantly it makes it easier to understand them and add new
ones, since you don't have to go up 600 lines to see which objects are
created by each tests, and thus which expectations to set.

* Explicitly make sure correct order is selected

* Random order, test & refactor

* Properly disable buttons

Through specific attribute and not only CSS.

* Link to full proposal when votes disabled