Rubocop improvements (#1451)
* Downgrade rubocop back to 0.48.1

Since it's the version codeclimate uses.

* Use rubocop's default included files

* Let rubocop enforce consistency everywhere

* Prefer inline exclude

So the rest of the file benefits from the cop.

* Normalize index creation

This is the only `using: :btree` usage. This is automatic.

* Enforce Style/AlignArray in specs too

* Enable Style/Lambda everywhere

* Unlock line length checker in specs too

* Remove some more unnecessary exclusions

* Disable Style/PredicateName inline when needed

* Ignore codeclimate duplication engine in specs

Too many issues at the moment.

* Simplify DB resetting

* Fix very subtle bug in coverage tracking

The root path was not properly expanded, resulting in things like
`/home/deivid/Code/decidim/decidim-api/..`. When this expression was
applied to the "external file filter" regexp, namely:

```
add_filter do |src|
  src.filename !~ /^#{root_path}/
end
```

the dots were interpreted as universal matchers, so the filter ended up
filtering out whatever was outside the plugin folder, not outside the
whole decidim repo.

This caused too many files to be filtered out from reports when the
covered code was in a different folder from the test code. For example,
the `spec/types/query_type_spec.rb` in `decidim-api` covers
`decidim-core/lib/decidim/query_extensions.rb` in `decidim-core`. This
coverage was being ignored.