Decouple global ability context
The global ability context shouldn't always try to look for the current
participatory process.

As a result, we can make the `current_participatory_process` method in
the participatory processes admin controller private. Public methods in
controllers should match the public controller route actions, so having
to have this method public was weird.