Final 404 handling normalization
There were two confusing routing statements that fell back to the 404
error action once no matching engine was found for a feature-like URL.
This was causing missing features to be different from other missing
content since in development you would get the real production 404 page
instead of a backtrace and error information.

The only reason these routes were required was to define the named
helpers `feature_path` and `manage_feature_path`.

So, to fix the issue, I removed the routes so that features behave just
like other missing helpers, and manually defined the named helpers to
give us the `root_path` for the feature in each engine.

The nice side effect is that now features _can_ define their own root
path, which may or may not be /process/:process_id/f/:feature_id.