Remove some usages of `allow_any_instance_of`
All the flaky test failures happen in specs using it, so my guess is
that removing it might fix the flakyness. The alternative approach I
chose is to test whatever the commentable supports, i.e., it they don't
support alignable or arguable comments, test that, instead of forcing
the commentable into supporting them via the `allow_any_instance_of`
stubbing. Note that this same approach is already used in this file to
check for notifications: it expects (or not) a notification to be
received via email, according to the commentable having (or not) an
author.

The way these tests look can certainly be improved to remove the
conditionals and split the negative and positive cases, but I'll leave
that for another time.