Fix/flaky comments tests (#1530)
* Simplify comments' tests

* s/notificatione/notifications/

* Prefer `create_list` to Array.new

Since it's the majoritary style.

* Reword comment specs to read better

* Remove some usages of `allow_any_instance_of`

All the flaky test failures happen in specs using it, so my guess is
that removing it might fix the flakyness. The alternative approach I
chose is to test whatever the commentable supports, i.e., it they don't
support alignable or arguable comments, test that, instead of forcing
the commentable into supporting them via the `allow_any_instance_of`
stubbing. Note that this same approach is already used in this file to
check for notifications: it expects (or not) a notification to be
received via email, according to the commentable having (or not) an
author.

The way these tests look can certainly be improved to remove the
conditionals and split the negative and positive cases, but I'll leave
that for another time.

* Extract common setup

* Remove assertion already checked by previous test

* Fix the last flaky failures

I observed that the only tests left randomly failing were the ones where
`login_as` was being used more than once. Since this method hacks into
the session instead of doing the whole login flow, I figured it could be
problematic to re-login without login out first.

Logging out before re-login seems to do the trick indeed.

* Address feedback