Fix weird menu reloading issue
In the development app, do this:

* Load http://localhost:3000/admin/participatory_processes/2/features/8/manage
* Remove letter_opener related routes (or anything else, I guess) from
  core routes file.
* Reload the page.

You get the error

```
A copy of Decidim::Admin::MenuHelper has been removed from the module
tree but is still active!
```

This seems like some bad interaction between ruby's built-in autoload
and Rails autoloading mechanism based on const_missing.

I don't fully understand it, but this fixes the issue.