Fix letter opener configuration (#1552)
* Better style for code snippets in generators

I think this approach gives us the best of both worlds:

* Nicely indented generating code.
* Nicely indented generated code.

And the target indentation is also easy to see.

* Fix letter_opener_web configuration

The letter_opener_web configuration was not working. When an email was
sent in development you would get an undefined method error, and the
route /letter_opener would give a 404.

I'm not sure what the issue was, but letter_web_opener is supposed to be
configured in the app side, not in the engine side. So I removed the
dependency from decidim-core and moved it to the generated app. Now it
works!