Refactor error handling & controller hooks (#1524)
* Don't require test app regeneration for `test_all`

Just like it's not required for the default task.

* Indentation fix

* Typo

* Move error pages to a separate controller

This allows us to return proper http status codes for the error
conditions.

* Improve one spec

Move internal setup inside context and external setup out of context.

* Extend controller only when needed

In the case of the participatory process index page, we are still
extending the controller with the `NeedsParticipatoryProcess` module.
This has several drawbacks:

* It includes a lot of methods that don't make sense for the index page,
like `current_participatory_process`.
* It requires hacks to "undo" the inclusion like the `skip_after_action`
hook in the controller.
* It makes the index action touch the DB unnecessarily. When authorizing
the action, the `ability_context` tries to call
`current_participatory_process`, and since the method is defined, DB is
touched. Things work "by chance" because the query returns `nil`, but
it's incorrect and inefficient.

This approach should fix these problems.

* Unnecessary include

* Fix another crash in route constraints

* Improve 404 errors

If we let them flow through our custom exception handling normally
instead of rescueing and rising custom exceptions, we get the same
behavior in production (a 404 page), but better error pages in
development, with code context & backtrace information, instead of a
single "Not found" message.

And simpler code, of course.

* Remove unnecessary namespacing

* Fail early on missing participatory process groups

Otherwise we fail when rendering the template. This is less efficient
since we do more unnecessary work, and it raises a different error
(`ActionView::Template::Error`) than the other "not found" cases. I'm
guessing Rails treats these just like `RoutingError` or `RecordNotFound`
error, but it feels better to be consistent.

* s/individial/individual

* Final 404 handling normalization

There were two confusing routing statements that fell back to the 404
error action once no matching engine was found for a feature-like URL.
This was causing missing features to be different from other missing
content since in development you would get the real production 404 page
instead of a backtrace and error information.

The only reason these routes were required was to define the named
helpers `feature_path` and `manage_feature_path`.

So, to fix the issue, I removed the routes so that features behave just
like other missing helpers, and manually defined the named helpers to
give us the `root_path` for the feature in each engine.

The nice side effect is that now features _can_ define their own root
path, which may or may not be /process/:process_id/f/:feature_id.

* Remove unnecessary map

Organization locales are already strings.

* Visually separate controller sections

* Redirect to system when there's no organization

The main layout needs an organization to be properly rendered, so our
error controller crashes in this cases. The easy way out seems to
redirect to /system since the case when this happens is likely to be a
configuration or setup problem anyways.

* Show a warning in /system when no organizations

* Skip authorization in ErrorsController

Trying to keep it as bare as possible, since errors here mess things up.