Prefer full constants to `require_dependency`
At some point I was specializing an `ApplicationController` under a
specific namespace. The `require_dependency
"decidim/admin/application_controller"` was actually forcing the global
`ApplicationController` to be used even when the actual constant
resolution should've pointed to my specialized class. That was
unexpected and hard to troubleshoot so I prefer the more explicity
approach of full namespacing.