Refactor/factories (#1636)
* Remove unused traits

* Remove unnecessary `after(:create)` hook

* Reuse trait

* Extract some logic to a variable

* Move test utility to a better place

* Normalize `sequence` usage

All the other instances specify no parentheses.

* Move `test_file` method to decidim-dev

Rather than defining it in the global namespace.

* Normalize factory requiring

Each plugin should only exposed its own factories. This revealed one
missing require in `decidim-budgets`.

* Remove unnecessary stuff from factories

I removed it because it was messing with tests I was writing, and it
turns out that it's not needed.

* Simplify image test setup

* Move user role factories to core

Since the class has been moved directly under `Decidim`.