Code reloading improvements (#1633)
* Prefer full constants to `require_dependency`

At some point I was specializing an `ApplicationController` under a
specific namespace. The `require_dependency
"decidim/admin/application_controller"` was actually forcing the global
`ApplicationController` to be used even when the actual constant
resolution should've pointed to my specialized class. That was
unexpected and hard to troubleshoot so I prefer the more explicity
approach of full namespacing.

* Fix another code reloading issue

This commit fixes another of those

```
A copy of <MyClassName> has been removed from the module tree but is
still active!
```

* Remove unnecessary `require_dependency` usages

These paths are not autoloaded anyways, so this does not make a
difference, I think.

* Normalize `extend`/`include` style

You normally don't use parenthesis when including or extending modules.

* Fix `NeedsOrganization` extension/inclusion

After removing most of the `require_dependency` calls for the different
`ApplicationController`'s, some tests started failing like this:

```
rspec ./spec/features/admin_manages_participatory_processes_spec.rb

undefined local variable or method `current_organization'
```

That reveals an actual bug in `NeedsOrganization`. We need to use plain
ruby for this module, since it is used via `extend` (not only via
`include`), so `ActiveSupport::Concern` doesn't suffice.

Not really sure why we haven't hit this in real life.