Fix bug when re-seeding database
To reproduce, do

* Run `bundle exec development_app`.
* Switch to the generated app, `cd development_app`.
* Edit `config/initializers/decidim.rb` by uncommenting the
  `config.available_locales` line.
* Run `bundle exec rails db:drop db:create db:migrate db:seed`.

You will get

```
rails aborted!
ActiveModel::UnknownAttributeError: unknown attribute 'verified_at' for
Decidim::UserGroup.
```

This is because one of the migrations loads the `Decidim::UserGroup`
model, _before_ the `verified_at` column has been added to the table. So
when we want to use the model while seeding, we have an incorrect
version of it.

I fixed it by using a dummy model in the migration instead of the real
class, and using raw SQL for updating each element.