Less participatory process coupling in core (#1652)
* Remove unnecessary custom exception

The query should never be instantiated with a nil argument, just like
any other query object shouldn't. I don't see a special reason to make
this one a special case.

* Split `OrganizationParticipatoryProcesses` query

The idea is to have a separate query class that only filters by
organization, and use that query instead of the `has_many` association
from the model. That way we remove the coupling between the
`Organization` model and the participatory process concept.

* Get rid of `participatory_processes` association

The organization model is not longer coupled with participatory
processes.

* No participatory process dependency on serializer

Proposal serializer is now participatory space agnostic.

* Remove `participatory_process_groups` association

Just like with `participatory_processes`, prefer a separate query
instead.