Finally fix flakyness? (#1685)
* Don't shallow load errors

I guess this was debugging info when this was setup, but it has the
effect of swallowing any load errors happening in the test app. Better
to let it raise.

* Simplify poltergeist extension loading

`__dir__` returns an absolute path, no need to expand it.

* Disable CSS3 and JQuery animations for tests

Apparently they prevent the comment form from being sent sometimes. I
can't tell flakyness in 100% fixed, but if it's not, it's still much
less frequent than before.