Remove model dependencies from migrations (#1745)
* Fix some model dependent migrations

* We no longer need to reload the whole Rails env

The reason for this was that the wrong version of models was being
loaded because migrations had previously used their "half-baked"
versions. This is now fixed so we no longer need to reload everything.

This should speed up development application creation and also make it
easier to detect this kind of little issues in the future.