Global locale checking (#1740)
* Remove duplicated i18n key

* Add missing translation

* Remove unnecessary `ignore_unused` exception

* Normalize some i18n calls

Our convention is to use the engine as the scope. Aditionally, i18n-tasks
is able to understand this form.

* Remove commented out `ignore_unused` exclusions

* Sort unused exclusions

* Remove unused i18n keys

* Inline helper method

I decided to remove this helper method because:
* It was only used once.
* It used relative keys, which is a bit unreadable inside a helper
  (relative to what?).
* It had an unused parameter making it more cryptic.

Inlining it made apparent another unused key.

* Fix incorrect locales loading in `check_locales`

Rails environment was not being loaded, so the set of locales was the
default.

* Centralize i18n-tasks config

This has several benefits:

* Single place for configuring i18n checks.
* Smarter "unused checks" because we now have context. This means a lot
less false positives on "unused translations" when the usage happens
in a different plugin from the definition. That means less exceptions in
configuration.
* Smarter "missing checks" because we now have context.

* Normalize account message keys

It was weird and it was causing strange issues.

* Normalize YAML array indentation

* Fix double comment symbols

* Bump i18n-tasks

* Remove more duplicated i18n keys

* Normalize i18n files

After all the dance in this PR, the normalize spec started failing. This
change fixes it. Hopefully Crowdin can live with this.

* Move error messages to a single place