Completely remove after login redirection stuff
I reviewed this and it's actually unused since what we do is render a
login modal via JS, so we don't really exercise the whole "save the
redirection URL in the session, and then redirect to it after login"
dance.

I think it's better to remove it for now since it might fix some weird
redirection bugs we are having?

It also fixes an UI bug where we have a link that says "take a look at
the current processes", but takes to an undeterministic location.