Don't swallow errors in app generation
The `rails_command` method shells out the target command, but swallows
any errors that may happen. In my opinion, this is bad because we really
want these commands to succeed, and since the output is usually long,
you might not notice that they have failed. In addition, if errors are
swallowed, it's really hard to unit test generators!