Merge demo generator with regular generator
Transforming an existing application like this is very prone to errors
and not that simple. For example, if there are existing authorization
handlers already saved in DB (in either organizations or
authorizations), they would need to be migrated or we would get crashes
all around.

So the current version of this generator is only ready to be used for
direct app generator. So I merged it with the regular generator and made
it accessible by passing the --demo flag to it.