Let each gem keep its own version
This is the only way to ensure that gems are really isolated and that we
don't depend on a "inter-gem" specific directory structure. Both rspec &
rails do the same thing, and I can't really see another way to get this
working without having to ship all gems inside the main `decidim` gem
like we are doing now. Even if that case, it's very prone to errors and
unflexibly anyways.