Don't use models in migrations (#1921)
* Run seeds in the same process as migrations

So we can detect bad model usage inside migrations.

* Don't swallow errors in app generation

The `rails_command` method shells out the target command, but swallows
any errors that may happen. In my opinion, this is bad because we really
want these commands to succeed, and since the output is usually long,
you might not notice that they have failed. In addition, if errors are
swallowed, it's really hard to unit test generators!

* Remove hardcoded models from migrations

And add a regression test for development app generation. To ensure that
these errors can be properly caught in the future.

* Prefer `abort` to `exit 1`

* Remove unnecessary `ActiveRecord::Base.connection`

In migrations, all that stuff in directly available.

* Silence test output