Bundle less generator failure (#1938)
* Fix crash when running decidim executable

We can't asume bundler is available inside our generators, since it
will not when they are run through decidim executable.

Fixes the following crash:

```
/code/to/decidim/lib/generators/decidim/install_generator.rb:25:in
`bundle_install': uninitialized constant
Decidim::Generators::InstallGenerator::Bundler (NameError)
```

* Improve some test descriptions

* Test generator without flags

* Extract a shared example

* Extract test app name to a variable

And rename it so that specs don't fail on a freshly cloned repo (where
./tmp does not exist).

* Allow DB dropping to fail

For example, on freshly cloned repos. As a result, we no longer need to
check for CI here.

* Generate dev application in tests via bin/decidim

So we can give it the standard "test name" and it doesn't affect our
"real" development application.