Add bootsnap to the generated application (#1941)
* Add bootsnap to the generated application

In my case, this speeds up boot by ~33% (from 4.8s to 3.2s).

* Remove misleading Windows support bit

We can add Windows support in the future, but I don't think a line in
the Gemfile is not...

This also fixes an annoying bundler warning:

```
The dependency tzinfo-data (>= 0) will be unused by any of the platforms
Bundler is installing for. Bundler is installing for ruby but the
dependency is only for x86-mingw32, x86-mswin32, x64-mingw32, java. To
add those platforms to the bundle, run `bundle lock --add-platform
x86-mingw32 x86-mswin32 x64-mingw32 java`.
```