Prepare release 0.6.6 (#1967)
* Don't use models in migrations (#1921)

* Run seeds in the same process as migrations

So we can detect bad model usage inside migrations.

* Don't swallow errors in app generation

The `rails_command` method shells out the target command, but swallows
any errors that may happen. In my opinion, this is bad because we really
want these commands to succeed, and since the output is usually long,
you might not notice that they have failed. In addition, if errors are
swallowed, it's really hard to unit test generators!

* Remove hardcoded models from migrations

And add a regression test for development app generation. To ensure that
these errors can be properly caught in the future.

* Prefer `abort` to `exit 1`

* Remove unnecessary `ActiveRecord::Base.connection`

In migrations, all that stuff in directly available.

* Silence test output

* Fix app generators (#1931)

* Bundle less generator failure (#1938)

* Fix crash when running decidim executable

We can't asume bundler is available inside our generators, since it
will not when they are run through decidim executable.

Fixes the following crash:

```
/code/to/decidim/lib/generators/decidim/install_generator.rb:25:in
`bundle_install': uninitialized constant
Decidim::Generators::InstallGenerator::Bundler (NameError)
```

* Improve some test descriptions

* Test generator without flags

* Extract a shared example

* Extract test app name to a variable

And rename it so that specs don't fail on a freshly cloned repo (where
./tmp does not exist).

* Allow DB dropping to fail

For example, on freshly cloned repos. As a result, we no longer need to
check for CI here.

* Generate dev application in tests via bin/decidim

So we can give it the standard "test name" and it doesn't affect our
"real" development application.

* Centralize dependencies (#1958)

* Style fixes in template Gemfile

* Centralize dependencies to a single Gemfile

* Upgrade dependencies

* Require survey TOS field on admin (#1980)

* Require survey TOS field on admin

* Add CHANGELOG entry

* Fix puma dep

* Fix input fields for single locale organizations (#1983)

* Fix input fields for single locale organizations

* Add CHANGELOG entry

* Bump version to 0.6.6

* Downgrade graphiql-rails

* Update CHANGELOG

* Use latest docker image on CircleCI