Decouple and slim down decidim (#2031)
* Fix license specification

Otherwise `gem build decidim` warns:

```
WARNING:  license value 'AGPLv3' is invalid.  Use a license identifier
from http://spdx.org/licenses or 'Nonstandard' for a nonstandard
license. Did you mean 'AGPL-3.0', 'LGPLLR', 'NGPL'?
```

* Unfold `add_default_gemspec_properties`

These properties change very little, and this metadata shouldn't
necessary change in-sync for all gems.

* Centralize Rails dependency in core for now

For simplicity.

* Let each gem keep its own version

This is the only way to ensure that gems are really isolated and that we
don't depend on a "inter-gem" specific directory structure. Both rspec &
rails do the same thing, and I can't really see another way to get this
working without having to ship all gems inside the main `decidim` gem
like we are doing now. Even if that case, it's very prone to errors and
unflexibly anyways.

* Normalize local variable name passed to gemspec

* Remove optional faker dependency

* Add an `update_versions` tasks to manage versions

This is greatly inspired by Rails strategy.

* Ship a thin (160Kb) instead of a fat (7.9Mb) gem

We were shipping the whole repository inside the decidim gem, but it was
due to a bug in the file regexp. However, that bug was actually keeping
generators (mainly) working since we used to depend on the internal
inter-gem directory structure.

This has been fixed now so we can go and fix the regexp bug.

* Remove unnecessary codeclimate exclusions