Fix `bundle exec rake docker_development_app` (#2149)
* Remove unnecessary mkdir

WORKDIR creates it by default if it doesn't exist.

* Use relative paths after workdir

* Cache each install step for efficiency

* Prefer COPY to ADD

Since we're not using any of the ADD extra features (remote URL fetching
or archive extraction).

* Remove stuff already done by the generator

* Remove unnecessary env variable passed to command

I think this was necessary due to a Rails bug that caused the
asset:precompile task to unintentionally touch DB. But I don't think
that happens anymore.

* Prevent bundler warning

```
Your Gemfile lists the gem decidim (>= 0) more than once.
You should probably keep only one of them.
While it's not a problem now, it could cause errors if you change the
version of one of them later.
```

* Prefer `rails` for Rails commands

* Rename `path` to `docker_app_path`

All generators use `path` to refer to the local path for the decidim
gem, not to the path for the target app. The dummy app for tests uses a
`dummy_app_path` parameter for this, so let's use something similar.

Hopefully this reduces confusion when trying to understand generators.

* Remove unnecessary code

* Minor simplification in docker generator

* More modern flag for gem install

* Prefer `db:seed` rather than `db:setup`

Since we've already created the DB and migrated it by that time.

* Fix missing migration in dev docker app creation

We need to make sure assemblies is used. Otherwise we get the following
error:

```
rails aborted!
ActiveRecord::StatementInvalid: PG::UndefinedTable:
ERROR:  relation "decidim_assemblies" does not exist
```

* Remove unnecessary arg value

* Use docker-compose v3

Otherwise I get the following error when doing `docker-compose up`:

```
`require': cannot load such file -- rack/handler/-b0.0.0.0 (LoadError)
```