generate_test_app task no longer needs be exposed (#2184)
A while ago we changed the task meant to be used for testing by external
plugings. It's `bundle exec rake decidim:generate_external_test_app`
now. That means we no longer need to expose our internal task.

So I moved it inside main generators and renamed it to `test_app` to be
consistent with the `development_app` task.