Fix admins being prevented from impersonating (#2372)
* Move helpers to private section

* No deferred authorizations for impersonation

When creating a new impersonation, all authorization methods are
oferred. However, only direct authorization methods are expected to work
for impersonation. This commit fixes that by not offering deferred
authorization methods for impersonation.

Also added some renaming to make the code clearer in this regard.

* I always underestimate flakies

* Fix more flakies

The first `have_no_selector` assertion was wrong, since the selector is
not removed from the page, only hidden... That leads to the question,
why was it working in the first place?

* Fix last flaky

* Typo

* Fix session expiration simulation

`perform_enqueued_jobs` runs the async jobs inmediately, even if they've
been scheduled to be run in the future. That means that the
impersonation was being expired before actually being started!

These changes better represent the real workflow.

* Fix admins no longer being able to impersonate

If an admin went straight to the admin console after an impersonation
expired, it would lose the impersonation session without being able to
get it properly expired. And thus it would be prevented from
impersonating again.

This changes fixes it.