Deterministic seeds
Sometimes we get weird coverage fluctuations in PRs, see [1] for
example. This is because seeds will sometimes randomly create no
participatory spaces with scopes, and that means some lines of code will
not be covered.

This commit fixes that by making sure seeds always create participatory
spaces with associated scopes.

[1]: https://codecov.io/gh/decidim/decidim/pull/2554/src/decidim-proposals/lib/decidim/proposals/feature.rb