Enhancement in scopes helper (#2432)
* Using form helper to retrieve participatory space scope(if available) in scopes helper.

* Refactored current_space_scope helper. Added docs for scopes helpers. Fixed has_visible_scopes? helper.
Converted to a controller helper and renamed to current_participatory_space_scope. Added reference to it from forms (and to current_participatory_space too, as it would be expected).
The `has_visible_scopes?` helper logic was valid only for a one level scope hierarchy. Changed to show the scope when scopes are enabled and resource scope is not the participatory space scope.

* Added new context attributes to form specs.

* Added `current_participatory_space` alias helper to assemblies and participatory processes admin controllers. Modified code that uses it to avoid errors if not present.

* Removed current_participatory_space_scope and moved back to try to recover scope from current_participatory_space, if exists.

* Test fixed.