Fix gem build warnings (#2566)
* Bump gems

* Use semver friendly version specificators

For maximum flexibility and to reduce the chance of conflicts with end
applications. And to avoid `gem build` warnings like the following:

```
WARNING:  pessimistic dependency on active_link_to (~> 1.0.0) may be overly strict
  if active_link_to is semantically versioned, use:
    add_runtime_dependency 'active_link_to', '~> 1.0', '>= 1.0.0'
```

Took extra care with gems that we know don't follow strict semver.

* Fix missing homepage gem build warning

```
WARNING:  no homepage specified
WARNING:  See http://guides.rubygems.org/specification-reference/ for help
```

* Centralize `date_validator` dependency in core

It's only required by core, and for example, participatory_processes
uses it and does not include. This change normalizes that.

* Remove `rectify` dependency from `decidim-debates`

We have that centralized in core.

* Fix more gemspec warnings on `gem build`

Like the following:

```
WARNING:  description and summary are identical
WARNING:  See http://guides.rubygems.org/specification-reference/ for help
```

* Fix & normalize some summaries & descriptions

* Bound rubocop-rspec version

To keep it in sync with codeclimate.

* Fix `Lint/MissingCopEnableDirective` offense

* Add TODO note

* Remove "plugin" word usage