Fix generator specs
For some reason they were passing on CI, but they were failing locally,
at least for me. Now they are more robust. Fixing this also allowed us
to remove the weirdness of having to do `BUNDLE_GEMFILE=Gemfile bundle
install`, which was just a workaround for the underlying problem of not
resetting the bundler environment before running `bundle install` on the
generated application.

This is also an actual bug fix, without these changes I'm unable to
generate a development application from an external module with a
modified bundler environment.