Remove another unnecessary option
This was added to allow creating test applications from inside an
application but this is not actually necessary. The use case was
decidim-barcelona having the `decidim-debates` component as a local
folder and wanting to test it. The solution would be instead to generate
the test application inside the component folder and it should just
work.