Generator improvements for external modules (#2699)
* Semantic versioning specifiers in Gemfile

* Replace only decidim and decidim-dev

* Fix generator specs

For some reason they were passing on CI, but they were failing locally,
at least for me. Now they are more robust. Fixing this also allowed us
to remove the weirdness of having to do `BUNDLE_GEMFILE=Gemfile bundle
install`, which was just a workaround for the underlying problem of not
resetting the bundler environment before running `bundle install` on the
generated application.

This is also an actual bug fix, without these changes I'm unable to
generate a development application from an external module with a
modified bundler environment.

* Fix warnings on decidim-debates gemspec

* Unhide `install_all` & `uninstall_all` output

We want clean test runs, but if run directly, it's useful to see the
output.

* Smart Gemfile guessing for generator

* Get external test app generation tested

* Remove unnecessary dependencies

* Expose development_app task for external modules

Also, I'm introducing yet another `Metrics/BlockLength` offense.
Personally, I think this metric is not useful for us since we keep
getting offenses for it and just ignoring them. I propose to remove the
rule and manually watch out for excesive complexity.

* s/rake db:create/rails db:create/

* Fix gem name guessing in component manager

The main repo folder is not necessarily called "decidim". For example,
it's called "code" in the docker image. If the folder does not match
/decidim/, asume we are managing the root gem.

* Remove unnecessary generator option

Since decidim is currently fully dependent on postgreSQL.

* Remove another unnecessary option

This was added to allow creating test applications from inside an
application but this is not actually necessary. The use case was
decidim-barcelona having the `decidim-debates` component as a local
folder and wanting to test it. The solution would be instead to generate
the test application inside the component folder and it should just
work.

* Use exact same tasks for core & external component

* Remove unnecessary and broken spec