Refactor scope modules
We previously had two different scope mixins, `HasScope` and `Scopable`
that were practically doing the same thing. `HasScope` was included by
components whereas `Scopable` was included by participatory spaces. The
only difference between them is that participatory spaces include a
boolean flag to define whether scopes are enabled or not. However, I
noticed that components also use that `scopes_enabled` property but
delegate it to the participatory space.

This commit unifies all that.