Refactor scopes (#2895)
* Refactor scope modules

We previously had two different scope mixins, `HasScope` and `Scopable`
that were practically doing the same thing. `HasScope` was included by
components whereas `Scopable` was included by participatory spaces. The
only difference between them is that participatory spaces include a
boolean flag to define whether scopes are enabled or not. However, I
noticed that components also use that `scopes_enabled` property but
delegate it to the participatory space.

This commit unifies all that.

* Extract an out_of_scope? method

* Remove unnecessary safe operators

* Remove unnecessary early returns

* Make custom validator a private method

* Move custom validations to private methods

* Remove duplicated relationship

* Visually separate file sections

* Grab scopes through participatory_space directly

* Remove artificial complexity

* Remove unnecessary defensiveness

* Fix proposals scope not displayed on process group

And remove unnecessary defensiveness while on it.