Fix unintentional swapping of answer options
Answer options, which cannot currently be reordered, use the same
component as questions for dynamic fields. In their case, they're not
supposed to have any "move up" or "move down" behaviour. However, they
were still binding the events, even though target selectors for "move
up" or "move down" were not being provided. That was making the events
to be bound globally and causing funny option swapping when focusing on
the input fields for filling them.

The regression test is admittedly cumbersome, sorry.