Stop using jquery-tmpl
The placeholder variable "${tabsId}" that is interpolated by this
library is causing errors when you submit a form with a freshly created
translated field and it has errors. The field will be rerendered with
the "${tabsId}" placeholder in the field tabs link destinations and that
will make foundation-tabs crash. That will make it impossible to further
switch tabs in that field.

This library has not been updated for 4 years anyways and new
alternatives are now recommended.