Fix broken translated field after form errors (#3026)
* Tiny refactor

* Dummy question form instead of dummy question

So the template can assume its locals are consistently the same type.

* Extract a before block

* Add a sad path spec for survey edition

* Change attribute naming to ease further refactors

* Another tiny refactor

Reuse the logic to construct the base id for the attributes' id
construction.

* Pass base id to the question template

So "calculation" is straightforward and we don't need helper methods.

* Reuse passed id for the wrapper container id

* Don't unnecessarily set id when adding new field

It'll have already been replaced with a proper value by then.

* Remove unnecessary check

* Normalize answer options check with questions

* Should be answer option, actually

* Stop passing a question to answer_option template

* Remove unnecessary `each_with_index`

* Stop using jquery-tmpl

The placeholder variable "${tabsId}" that is interpolated by this
library is causing errors when you submit a form with a freshly created
translated field and it has errors. The field will be rerendered with
the "${tabsId}" placeholder in the field tabs link destinations and that
will make foundation-tabs crash. That will make it impossible to further
switch tabs in that field.

This library has not been updated for 4 years anyways and new
alternatives are now recommended.

* Extract functions to try make code simpler

* Fix translated field not working after form errors

* Remove unnecessary line

* Don't replace globally but only attributes needed